requirements.txt
alembic==1.15.2
apispec==6.8.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Migrate==4.1.0
flask-smorest==0.46.0
Flask-SQLAlchemy==3.1.1
gitdb==4.0.12
GitPython==3.1.41
greenlet==3.2.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
packaging==25.0
pluggy==1.5.0
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg-pool==3.2.6
psycopg2==2.9.10
pytest==8.3.5
python-dotenv==1.1.0
setuptools==75.6.0
smmap==5.0.2
SQLAlchemy==2.0.40
typing_extensions==4.13.2
webargs==8.7.0
Werkzeug==3.1.3
Flask-JWT-Extended>=4.4.4
flask-cors>=4.0.0
